backend/requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
httplib2==0.13.0
idna==2.8
mysqlclient==1.4.2.post1
oauthlib==3.0.2
pkg-resources==0.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
